+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 1.1.0 (2012-XX-XX)
 
+* Added a recursive filter iterator keeping only displayed items
 * Added a filter iterator keeping only current items
-* Made the iterator recursive for the item
+* Added a recursive iterator for the item
 * Fixed building an array of breadcrumbs when a label has only digits
 * Added a way to mark a label as safe
 * Refactored the ListRenderer to be consistent with the TwigRenderer and provide the same extension points